For redesign
The for loop is no longer used as "forEach"
instead it's now used with a regular for loop
and it'll check if the function/string it comes across
is defined or undefined.

If it's undefined then it'll define it by the accpro function.
Else it will just move on.

the "Remove" method is now made to where it'll access the parent and
remove the element that's in the parent.
It will still return the element