Selectors Downgrade
Downgraded the selectors why is because, they were causing an error to
where. "Html of Null" meaning that if the selectors returned a null
property the formatter.

(that's the for loop in pendef)

Would break, the formatter right now is not the best and so is the
library but will be worked on in the future.