Edited website
Made the website a bit more responsive hopefully.

The updates are to the application and the style.

Also I removed the semi-colons, yes I know about the warnings
but frankly I don't care.

I have my reasons and you have yours let's leave it at that.

But hopefully the updates will make the webpage more responsive.

Will add more things in the future.