Experimental!
added an experimental update only,
to where in  "Advanced" the main parser will see if there's attributes
if there's attributes it'll automatically set the id and stuff
this is an advanced use only meaning in the base it won't be in there.
So make sure you download the right one.