Base and refactor
so base had some of the code from advanced.
buuut now it's gone and it's literally the base.

Also <penInstance>.html now takes another argument
parse - will parse as html
basically innerHTML adds a delay to load time
this delay is caused by the html parser
so by default it'll assume that you're referencing "textContent"
instead of "innerHTML" but if you set this option to true
then it'll use "innerHTML" and not "textContent"