Lol
So about the variables I was joking. Not actually going to do that.
Just really meant it when I said I made namespaces
It's not recursive so it only does it with top level names
like background, animation, text etc.

Also I now noticed that in the code I set the parent to be undefined
which was my bad when removing the parent so that bug is fixed.

I really didn't realize it till now, but now I did so there you go.