Finder
I now created a way to loop through an object.

and find a key with a value
but if it doesn't find it.

Then return a value provided in the function call
pen.findInObj(randomObj, 'someRandomKeyName', value)

which possibly this is already a feature but I don't know about it.

myeh